feat: moved config parsing to config.lua
